+++ integrations/feishu-bridge/src/lib.mjs
@@ -1,60 +1,37 @@
-export function parseList(raw) {
-  return String(raw || "")
-    .split(",")
-    .map((item) => item.trim())
-    .filter(Boolean);
-}
+import {
+  activeTurnBlock,
+  cleanEnvValue,
+  commandAction as coreCommandAction,
+  compactRuntimeError,
+  isPlaceholderValue,
+  latestRunningTurn,
+  parseApprovalDecisionArgs,
+  parseBool,
+  parseCommand,
+  parseEnvText,
+  parseList,
+  parseTextContent as coreParseTextContent,
+  preservedChatStateFields as corePreservedChatStateFields,
+  splitMessage,
+  stripGroupPrefix as coreStripGroupPrefix
+} from "../../bridge-core/src/lib.mjs";
 
-export function parseBool(raw, fallback = false) {
-  if (raw == null || raw === "") return fallback;
-  return ["1", "true", "yes", "on"].includes(String(raw).trim().toLowerCase());
-}
-
-export function parseEnvText(raw) {
-  const env = {};
-  for (const line of String(raw || "").split(/\r?\n/)) {
-    const trimmed = line.trim();
-    if (!trimmed || trimmed.startsWith("#")) continue;
-    const normalized = trimmed.startsWith("export ") ? trimmed.slice(7).trim() : trimmed;
-    const index = normalized.indexOf("=");
-    if (index <= 0) continue;
-    const key = normalized.slice(0, index).trim();
-    let value = normalized.slice(index + 1).trim();
-    if (
-      (value.startsWith('"') && value.endsWith('"')) ||
-      (value.startsWith("'") && value.endsWith("'"))
-    ) {
-      value = value.slice(1, -1);
-    }
-    env[key] = value;
-  }
-  return env;
-}
-
-export function cleanEnvValue(value) {
-  return String(value ?? "").trim();
-}
-
-export function isPlaceholderValue(value) {
-  const normalized = cleanEnvValue(value).toLowerCase();
-  return (
-    !normalized ||
-    normalized.includes("replace-with") ||
-    normalized.includes("xxxxxxxx") ||
-    normalized === "changeme"
-  );
-}
+export {
+  activeTurnBlock,
+  cleanEnvValue,
+  compactRuntimeError,
+  isPlaceholderValue,
+  latestRunningTurn,
+  parseApprovalDecisionArgs,
+  parseBool,
+  parseCommand,
+  parseEnvText,
+  parseList,
+  splitMessage
+};
 
 export function parseTextContent(content) {
-  if (typeof content !== "string") return "";
-  try {
-    const parsed = JSON.parse(content);
-    if (typeof parsed.text === "string") return parsed.text;
-    if (typeof parsed.content === "string") return parsed.content;
-  } catch {
-    return content;
-  }
-  return content;
+  return coreParseTextContent(content, ["text", "content"]);
 }
 
 export function incomingIdentity(event) {
@@ -98,124 +75,20 @@ export function pairingRefusalText(identity) {
 }
 
 export function stripGroupPrefix(text, { chatType, requirePrefix, prefix }) {
-  const trimmed = String(text || "").trim();
-  if (!trimmed) return { accepted: false, text: "" };
-  if (!requirePrefix || chatType === "p2p") {
-    return { accepted: true, text: trimmed };
-  }
-  const marker = prefix || "/ds";
-  if (trimmed === marker) return { accepted: true, text: "/help" };
-  if (trimmed.startsWith(`${marker} `)) {
-    return { accepted: true, text: trimmed.slice(marker.length).trim() };
-  }
-  return { accepted: false, text: "" };
-}
-
-export function parseCommand(text) {
-  const trimmed = String(text || "").trim();
-  if (!trimmed.startsWith("/")) return { name: "prompt", args: trimmed };
-  const [head, ...rest] = trimmed.split(/\s+/);
-  return {
-    name: head.slice(1).toLowerCase(),
-    args: rest.join(" ").trim()
-  };
-}
-
-export function parseApprovalDecisionArgs(args) {
-  const parts = String(args || "")
-    .split(/\s+/)
-    .filter(Boolean);
-  return {
-    approvalId: parts[0] || "",
-    remember: parts.slice(1).includes("remember")
-  };
+  return coreStripGroupPrefix(text, {
+    chatType,
+    requirePrefix,
+    prefix: prefix || "/ds",
+    directChatTypes: ["p2p"]
+  });
 }
 
 export function commandAction(command) {
-  switch (command.name) {
-    case "help":
-      return { kind: "help" };
-    case "status":
-      return { kind: "status" };
-    case "threads":
-      return { kind: "threads" };
-    case "new":
-      return { kind: "new_thread" };
-    case "resume":
-      return { kind: "resume", threadId: command.args };
-    case "interrupt":
-      return { kind: "interrupt" };
-    case "compact":
-      return { kind: "compact" };
-    case "model":
-      // /model <model_name> — switch per-chat default model.
-      // Stored in thread store and used for future threads/turns.
-      // Pass "default" to reset to the bridge-level default.
-      return { kind: "set_model", modelName: command.args };
-    case "allow":
-      return { kind: "approval", decision: "allow", ...parseApprovalDecisionArgs(command.args) };
-    case "deny":
-      return { kind: "approval", decision: "deny", ...parseApprovalDecisionArgs(command.args) };
-    case "prompt":
-      return { kind: "prompt", prompt: command.args };
-    default:
-      return {
-        kind: "prompt",
-        prompt: `/${command.name}${command.args ? ` ${command.args}` : ""}`
-      };
-  }
+  return coreCommandAction(command);
 }
 
 export function preservedChatStateFields(state = {}) {
-  const preserved = {};
-  if (Object.prototype.hasOwnProperty.call(state || {}, "model")) {
-    preserved.model = state.model || null;
-  }
-  if (state?.replyToMessageId) {
-    preserved.replyToMessageId = state.replyToMessageId;
-  }
-  return preserved;
-}
-
-export function splitMessage(text, maxChars = 3500) {
-  const value = String(text || "");
-  const chars = Array.from(value);
-  if (chars.length <= maxChars) return value ? [value] : [];
-  const chunks = [];
-  let cursor = 0;
-  while (cursor < chars.length) {
-    chunks.push(chars.slice(cursor, cursor + maxChars).join(""));
-    cursor += maxChars;
-  }
-  return chunks;
-}
-
-export function compactRuntimeError(status, body) {
-  const message =
-    body?.error?.message ||
-    body?.message ||
-    (typeof body === "string" ? body : JSON.stringify(body));
-  return `Runtime API request failed (${status}): ${message}`;
-}
-
-export function latestRunningTurn(detail) {
-  const turns = Array.isArray(detail?.turns) ? detail.turns : [];
-  for (let index = turns.length - 1; index >= 0; index -= 1) {
-    const turn = turns[index];
-    if (["queued", "in_progress"].includes(turn?.status)) return turn;
-  }
-  return null;
-}
-
-export function activeTurnBlock(detail, state = {}) {
-  const runningTurn = latestRunningTurn(detail);
-  if (!runningTurn) return null;
-  return {
-    turnId: runningTurn.id || state.activeTurnId || "",
-    message: `Thread already has active turn ${
-      runningTurn.id || state.activeTurnId || "(unknown)"
-    }. Wait for it to finish or send /interrupt.`
-  };
+  return corePreservedChatStateFields(state, ["model", "replyToMessageId"]);
 }
 
 export function validateBridgeConfig(env, options = {}) {
