+++ integrations/weixin-bridge/src/lib.mjs
@@ -1,5 +1,19 @@
 import crypto from "node:crypto";
 
+import {
+  activeTurnBlock,
+  commandAction,
+  compactRuntimeError,
+  envFirst,
+  latestRunningTurn,
+  parseApprovalDecisionArgs,
+  parseBool,
+  parseCommand,
+  parseList,
+  preservedChatStateFields,
+  splitMessage,
+} from "../../bridge-core/src/lib.mjs";
+
 // ============================================================================
 // iLink Bot API 协议层 — 参考 @tencent-weixin/openclaw-weixin
 // ============================================================================
@@ -7,29 +21,19 @@ import crypto from "node:crypto";
 const DEFAULT_API_TIMEOUT_MS = 30_000;
 const LONGPOLL_DEFAULT_TIMEOUT_MS = 35_000;
 
-// ---------------------------------------------------------------------------
-// 通用工具
-// ---------------------------------------------------------------------------
-
-export function parseList(raw) {
-  return String(raw || "")
-    .split(",")
-    .map((item) => item.trim())
-    .filter(Boolean);
-}
-
-export function parseBool(raw, fallback = false) {
-  if (raw == null || raw === "") return fallback;
-  return ["1", "true", "yes", "on"].includes(String(raw).trim().toLowerCase());
-}
-
-export function envFirst(env, ...names) {
-  for (const name of names) {
-    const value = env?.[name];
-    if (value != null && String(value).trim()) return String(value).trim();
-  }
-  return "";
-}
+export {
+  activeTurnBlock,
+  commandAction,
+  compactRuntimeError,
+  envFirst,
+  latestRunningTurn,
+  parseApprovalDecisionArgs,
+  parseBool,
+  parseCommand,
+  parseList,
+  preservedChatStateFields,
+  splitMessage,
+};
 
 export function randomUin() {
   const uint32 = crypto.randomBytes(4).readUInt32BE(0);
@@ -72,125 +76,6 @@ export function extractText(itemList) {
     }
   }
   return "";
-}
-
-// ---------------------------------------------------------------------------
-// 命令解析（与 feishu/telegram/wechat bridge 一致）
-// ---------------------------------------------------------------------------
-
-export function parseCommand(text) {
-  const trimmed = String(text || "").trim();
-  if (!trimmed.startsWith("/")) return { name: "prompt", args: trimmed };
-  const [head, ...rest] = trimmed.split(/\s+/);
-  return {
-    name: head.slice(1).toLowerCase(),
-    args: rest.join(" ").trim(),
-  };
-}
-
-export function parseApprovalDecisionArgs(args) {
-  const parts = String(args || "").split(/\s+/).filter(Boolean);
-  return {
-    approvalId: parts[0] || "",
-    remember: parts.slice(1).includes("remember"),
-  };
-}
-
-export function commandAction(command) {
-  switch (command.name) {
-    case "help":
-      return { kind: "help" };
-    case "status":
-      return { kind: "status" };
-    case "threads":
-      return { kind: "threads" };
-    case "new":
-      return { kind: "new_thread" };
-    case "resume":
-      return { kind: "resume", threadId: command.args };
-    case "model":
-      return { kind: "set_model", modelName: command.args };
-    case "interrupt":
-      return { kind: "interrupt" };
-    case "compact":
-      return { kind: "compact" };
-    case "allow":
-      return {
-        kind: "approval",
-        decision: "allow",
-        ...parseApprovalDecisionArgs(command.args),
-      };
-    case "deny":
-      return {
-        kind: "approval",
-        decision: "deny",
-        ...parseApprovalDecisionArgs(command.args),
-      };
-    case "prompt":
-      return { kind: "prompt", prompt: command.args };
-    default:
-      return {
-        kind: "prompt",
-        prompt: `/${command.name}${command.args ? ` ${command.args}` : ""}`,
-      };
-  }
-}
-
-export function preservedChatStateFields(state = {}) {
-  const preserved = {};
-  if (Object.prototype.hasOwnProperty.call(state || {}, "model")) {
-    preserved.model = state.model || null;
-  }
-  return preserved;
-}
-
-// ---------------------------------------------------------------------------
-// 消息拆分（微信 iLink 单条消息无明确上限，保守 3500 字符）
-// ---------------------------------------------------------------------------
-
-export function splitMessage(text, maxChars = 3500) {
-  const value = String(text || "");
-  const chars = Array.from(value);
-  if (chars.length <= maxChars) return value ? [value] : [];
-  const chunks = [];
-  let cursor = 0;
-  while (cursor < chars.length) {
-    chunks.push(chars.slice(cursor, cursor + maxChars).join(""));
-    cursor += maxChars;
-  }
-  return chunks;
-}
-
-// ---------------------------------------------------------------------------
-// Runtime 工具
-// ---------------------------------------------------------------------------
-
-export function compactRuntimeError(status, body) {
-  const message =
-    body?.error?.message ||
-    body?.message ||
-    (typeof body === "string" ? body : JSON.stringify(body));
-  return `Runtime API request failed (${status}): ${message}`;
-}
-
-export function latestRunningTurn(detail) {
-  const turns = Array.isArray(detail?.turns) ? detail.turns : [];
-  for (let index = turns.length - 1; index >= 0; index -= 1) {
-    const turn = turns[index];
-    if (["queued", "in_progress"].includes(turn?.status)) return turn;
-  }
-  return null;
-}
-
-export function activeTurnBlock(detail, state = {}) {
-  const runningTurn = latestRunningTurn(detail);
-  if (!runningTurn) return null;
-  return {
-    turnId: runningTurn.id || state.activeTurnId || "",
-    message: `Thread already has active turn ${
-      runningTurn.id || state.activeTurnId || "(unknown)"
-    }. Wait for it to finish or send /interrupt.`,
-  };
 }
 
 // ---------------------------------------------------------------------------
